backend/requirements.txt
# FastAPI framework & ASGI server
fastapi
uvicorn[standard]

# Google Gemini API (Vision + Text)
google-generativeai

# Environment variables loader
python-dotenv

# Image handling (FastAPI-compatible uploads)
python-multipart

# Data validation / models (used by FastAPI)
pydantic

# LangChain core + Google Gemini integration (used in ai_notes)
langchain
langchain-google-genai

# MongoDB async driver
motor
pymongo

# Firebase ID token verification
firebase-admin

# Optional: helpful utilities
requests
